+++ CPM for OS X/BIOS/TerminalControlSet+ADM3A.m
@@ -36,8 +36,8 @@
 		{@"\x0c",	0,	^{	[weakSelf rightCursor];					}},
 		{@"\33=",	4,	^{
 							[weakSelf
-									setCursorX:(weakSelf.inputQueue[3] - 32)%weakSelf.width
-									y:(weakSelf.inputQueue[2] - 32)%weakSelf.height];
+									setCursorX:(NSUInteger)(weakSelf.inputQueue[3] - 32)%weakSelf.width
+									y:(NSUInteger)(weakSelf.inputQueue[2] - 32)%weakSelf.height];
 						}},
 
 		{@"\33B0",	0,	^{	weakSelf.currentAttribute |= kCPMTerminalAttributeInverseVideoOn;		}},
